// src/app/modules/main-container/models/reply.ts
export class Reply {
    messageId: number;
    groupTypeId: number;
    messageTypeId: number;
    contentId: number;
    parentContentId: number;
    content: string;
    attachments: string;
    receiverId: string;
    uniqueId: string;
    fullName: string;
    regUser: string;
    profileAddress: string;

}